fix(playground): move tool call config to a separate component
